Sınav çözümü

+++ src/com/banu/MaasOzelListe.java
@@ -2,16 +2,16 @@ package com.banu;
 
 public class MaasOzelListe {
 
-	int[] maasOzelListe;
+	double[] maasOzelListe;
 	int boyut = 0;
 
-	public void add(int maas) {
-		int[] tmp;
+	public void add(double maas) {
+		double[] tmp;
 		if (maasOzelListe == null) {
-			maasOzelListe = new int[1];
+			maasOzelListe = new double[1];
 			maasOzelListe[0] = maas;
 		} else {
-			tmp = new int[boyut + 1];
+			tmp = new double[boyut + 1];
 			for (int i = 0; i < boyut; i++) {
 				tmp[i] = maasOzelListe[i];
 			}
